+++ packages/@oak-tree-house/vuepress-theme-vuetify-blog/components/SidebarWordCloud.vue
@@ -1,29 +1,42 @@
 <template>
-  <v-container class="ma-1">
-    <div
-      class="font-weight-bold .subtitle-1 my-1"
-    >
-      {{ name }}
-    </div>
-    <div
-      class="d-flex flex-wrap"
-    >
-      <v-chip
-        v-for="word in words"
-        :key="word.name"
-        :to="word.path"
-        class="ma-1 word-chip"
-        outlined
-        color="primary"
+  <v-expansion-panels
+    v-model="value"
+    class="app-word-cloud"
+    accordion
+    flat
+    tile
+  >
+    <v-expansion-panel>
+      <v-expansion-panel-header
+        class="font-weight-bold .subtitle-1"
       >
-        {{ word.name }}
-      </v-chip>
-    </div>
-  </v-container>
+        {{ name }}
+      </v-expansion-panel-header>
+      <v-expansion-panel-content>
+        <div
+          class="d-flex flex-wrap"
+        >
+          <v-chip
+            v-for="word in words"
+            :key="word.name"
+            :to="word.path"
+            class="ma-1 word-chip"
+            outlined
+            color="primary"
+          >
+            {{ word.name }}
+          </v-chip>
+        </div>
+      </v-expansion-panel-content>
+    </v-expansion-panel>
+  </v-expansion-panels>
 </template>
 
 <script>
 export default {
+  data: () => ({
+    value: 0
+  }),
   computed: {
     name () {
       return this.$site.themeConfig.wordCloudName || 'Tag Cloud'
@@ -35,10 +48,14 @@ export default {
 }
 </script>
 
-<style lang="scss">
-.word-chip {
-  &:hover {
-    text-decoration: underline;
-  }
-}
+<style lang="sass">
+.app-word-cloud
+  .v-expansion-panel-header
+    padding: 0 16px
+    min-height: 48px !important
+  .v-expansion-panel-content__wrap
+    padding: 0 16px 16px
+  .word-chip
+    &:hover
+      text-decoration: underline
 </style>